iteration is int

+++ src/main/java/at/ngmpps/fjsstt/model/problem/Solution.java
@@ -251,7 +251,7 @@ public class Solution implements Serializable {
 	/**
 	 * @return the mIteration
 	 */
-	public long getIteration() {
+	public int getIteration() {
 		return iteration;
 	}
